extracted abstract base class for populations
